Sync a Folder between USB and Computer
This will sync a folder on a USB key with a path on your computer
specified below. It will take the most recently modified version of a
file and will sync that version between the USB key and the folder on
the computer.
This file should be run from a USB key
To do this for multiple folders just copy and paste all this code again
and set the new path and name for the other folder you want to backup